requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
ta==0.11.0
pyarrow==14.0.1
fastparquet==2023.10.1
pydantic==2.5.2
pydantic-settings==2.1.0
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
requests==2.31.0
websocket-client==1.7.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
python-telegram-bot==20.7
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
upstox-python-sdk==2.0.0
